backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyJWT==2.8.0
bcrypt==4.1.2
pytest==7.4.3
pytest-django==4.7.0
hypothesis==6.92.1
